+++ src/main/java/com/almostreliable/summoningrituals/altar/AltarBlockEntity.java
@@ -73,8 +73,6 @@ public class AltarBlockEntity extends BlockEntity implements TickableBlockEntity
     public long resetStartTick = -1;
     public long craftStartTick = -1;
     public float lastWaveAnchor;
-    public float orbitRotation;
-    public long orbitLastNanos = -1;
 
     public AltarBlockEntity(BlockPos pos, BlockState state) {
         super(Registration.ALTAR_BLOCK_ENTITY.get(), pos, state);
@@ -347,17 +345,17 @@ public class AltarBlockEntity extends BlockEntity implements TickableBlockEntity
         return currentRecipeInfo;
     }
 
-    @OnlyIn(Dist.CLIENT)
-    public void receiveRecipeInfoFromServer(@Nullable RecipeInfo currentRecipeInfo) {
-        this.currentRecipeInfo = currentRecipeInfo;
-    }
-
     public int getRecipeProgress() {
         return recipeProgress;
     }
 
     public int getRecipeTime() {
         return recipeTime;
+    }
+
+    @OnlyIn(Dist.CLIENT)
+    public void receiveRecipeInfoFromServer(@Nullable RecipeInfo currentRecipeInfo) {
+        this.currentRecipeInfo = currentRecipeInfo;
     }
 
     @OnlyIn(Dist.CLIENT)