+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
-- /
+
+- changed animation calculation base from system nano time to Minecraft's DeltaTracker
 
 ## [3.13.1] - 2026-04-28